Add JobGroup page and move the logic out from AllJobsPage so that it can be shared.
